requirements.txt
selenium
webdriver-manager
pytest
pytest-order
openpyxl
allure-pytest
